+++ InvoiceGenerator.Backend/InvoiceGenerator.Backend.BatchService/IBatchService.cs
@@ -13,15 +13,5 @@ namespace InvoiceGenerator.Backend.BatchService
         Task<ProcessingStatus> GetBatchInvoiceProcessingStatus(Guid processBatchKey, CancellationToken cancellationToken = default);
  
         Task<InvoiceData> GetIssuedInvoice(string invoiceNumber, CancellationToken cancellationToken = default);
-
-        Task<IEnumerable<InvoiceTemplateInfo>> GetInvoiceTemplates(CancellationToken cancellationToken = default);
-
-        Task<InvoiceTemplateData> GetInvoiceTemplate(Guid templateId, CancellationToken cancellationToken = default);
-
-        Task RemoveInvoiceTemplate(Guid templateId, CancellationToken cancellationToken = default);
-
-        Task ReplaceInvoiceTemplate(Guid templateId, InvoiceTemplateData invoiceTemplateData, CancellationToken cancellationToken = default);
-
-        Task<Guid> AddInvoiceTemplate(InvoiceTemplate invoiceTemplate, CancellationToken cancellationToken = default);
     }
 }
