+++ InvoiceGenerator.Backend/InvoiceGenerator.Backend.Cqrs/InvoiceGenerator.Backend.Cqrs.csproj
@@ -8,6 +8,7 @@
       <ProjectReference Include="..\InvoiceGenerator.Backend.Core\InvoiceGenerator.Backend.Core.csproj" />
       <ProjectReference Include="..\InvoiceGenerator.Backend.BatchService\InvoiceGenerator.Backend.BatchService.csproj" />
       <ProjectReference Include="..\InvoiceGenerator.Backend.Shared\InvoiceGenerator.Backend.Shared.csproj" />
+      <ProjectReference Include="..\InvoiceGenerator.Backend.TemplateService\InvoiceGenerator.Backend.TemplateService.csproj" />
       <ProjectReference Include="..\InvoiceGenerator.Backend.UserService\InvoiceGenerator.Backend.UserService.csproj" />
       <ProjectReference Include="..\InvoiceGenerator.Backend.VatService\InvoiceGenerator.Backend.VatService.csproj" />
     </ItemGroup>
