+++ InvoiceGenerator.Backend/InvoiceGenerator.Backend.TemplateService/Models/FileResult.cs
@@ -1,0 +1,12 @@
+namespace InvoiceGenerator.Backend.TemplateService.Models
+{
+    using System.Diagnostics.CodeAnalysis;
+
+    [ExcludeFromCodeCoverage]
+    public class FileResult
+    {
+        public byte[] ContentData { get; set; }
+
+        public string ContentType { get; set; }
+    }
+}
